+++ src/main/kotlin/com/home/autowatering/model/business/PotState.kt
@@ -6,6 +6,6 @@ data class PotState(
     var id: Long? = null,
     val pot: Pot,
     val date: Date,
-    val humidity: Double,
+    val humidity: Int,
     val watering: Boolean? = false
 )
